requirements/base.txt
# Core dependencies
mcp>=1.9.0
aiohttp>=3.9.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
openinference-instrumentation-anthropic>=0.1.0
openinference-semantic-conventions>=0.1.0
pyyaml>=6.0.0
anthropic>=0.5.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=2.0.0
loguru>=0.7.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
